requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-classic
chromadb
python-dotenv
openai
